A4A Dashboard: Address review feedback on the Team data layer
- Split the non-GET agency-team endpoints into a dedicated mutators.ts,
  keeping fetchers.ts for GET only (api-core convention).
- Normalize team member/invite data inside queryFn instead of select, so the
  mapping also applies when preloading via ensureQueryData.
- Use createInterpolateElement without sprintf in the team action modal.
